feat(api): move autocompleteBlockHash function to optimism.ts and remove old file
